requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==3.2.13
django-braces==1.15.0
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.1.1
pkg_resources==0.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
whitenoise==6.2.0
